ATLP-535 - move toDTO to controller
