+++ src/main/java/org/ohdsi/webapi/job/NotificationService.java
@@ -1,7 +1,9 @@
 package org.ohdsi.webapi.job;
 
+import org.springframework.batch.core.JobExecution;
+
 import java.util.List;
 
 public interface NotificationService {
-    List<JobExecutionResource> findAll();
+    List<JobExecution> findAll();
 }
